use better method and class names
